BOL-375 refactor: auth VM package 이동 (viewmodel.login -> view.login.auth)
