+++ presentation/src/main/java/com/yapp/bol/presentation/view/login/auth/KakaoTestActivity.kt
@@ -13,8 +13,6 @@ import com.yapp.bol.presentation.base.BaseActivity
 import com.yapp.bol.presentation.databinding.ActivityKakaoTestBinding
 import com.yapp.bol.presentation.utils.collectWithLifecycle
 import com.yapp.bol.presentation.view.login.LoginActivity
-import com.yapp.bol.presentation.viewmodel.login.AuthViewModel
-import com.yapp.bol.presentation.viewmodel.login.LoginType
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.flow.filterNotNull
