+++ presentation/src/main/java/com/yapp/bol/presentation/view/login/auth/NaverTestActivity.kt
@@ -11,8 +11,6 @@ import com.yapp.bol.presentation.R
 import com.yapp.bol.presentation.utils.collectWithLifecycle
 import com.yapp.bol.presentation.utils.showToast
 import com.yapp.bol.presentation.view.login.LoginActivity
-import com.yapp.bol.presentation.viewmodel.login.AuthViewModel
-import com.yapp.bol.presentation.viewmodel.login.LoginType
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.flow.filterNotNull
